requirements.txt
APScheduler==3.6.3
argh==0.26.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
cryptography==36.0.1
cycler==0.11.0
fonttools==4.29.1
idna==3.3
injector==0.19.0
kiwisolver==1.3.2
matplotlib==3.5.1
mysqlclient==2.1.0
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
python-slugify==5.0.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywaffle==0.6.4
PyYAML==6.0
requests==2.26.0
six==1.16.0
text-unidecode==1.3
tornado==6.1
tzdata==2021.5
tzlocal==4.1
unicode==2.8
urllib3==1.26.7
waffle==0.5.0
